refine copy hd op
